requirements.txt
#Heroku
Django>=1.7
gunicorn>=19.3.0
psycopg2==2.5.4
dj-database-url==0.3.0
dj-static==0.0.6
wsgiref==0.1.2
gevent>=1.0.1
celery>=3.1.15
django-celery>=3.1.16
django-picklefield
newrelic

#Site Specific
django-taggit==0.12.2
django_compressor>=1.3
jsmin>=2.0.6
stripe==1.9.8
django-timezone-field
Markdown>=2.0.3
Pillow>=1.7.7
#django-zebra
djangorestframework
mixpanel-py
python-postmark>=0.4.3
reportlab==2.7
xhtml2pdf>=0.0.5

#caching
django-redis-cache

#Search
pyelasticsearch==0.6.1
django-haystack==2.1.0

#Storage
boto==2.9.6
django-queued-storage>=0.6
django-storages==1.1.8

#Scraping
beautifulsoup4
